requirements.txt
fastapi
uvicorn
beanie
motor
pydantic-settings
python-dotenv
